ConfigurationService added. This service is launched at the default port and is used to get / set the information to the configuration file.
Data contracts added

Configuration changed, now it contains CurrentDevice which has port, host, name, and the list of active services.

When configuration is loaded, and the host and device name are empty, they are pre-filled with the host name and the machine name.

Some other minor corrections of the namespaces, etc.